% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getK.R
\name{getK}
\alias{getK}
\title{Estimate the number of common factors K}
\usage{
getK(y, X, Kmax = 12)
}
\arguments{
\item{y}{Response, T by 1 vector}

\item{X}{Predictors, p by T matrix}

\item{Kmax}{A prescribed upper bound that possibly increases with p and T
(default = 12)}
}
\value{
Estimate of K
}
\description{
Estimate the number of common factors K
}
\references{
Bai, J., and Ng, S. (2002), Determining the number of factors in approximate
factor models, \emph{Econometrica} 70(1), 191-221.

Li, H., Li, Q. and Shi, Y. (2017), Determining the number of factors when the
number of factors can increase with sample size, \emph{Journal of Econometrics}
197(1), 76–86.
}
