% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sufficientForecasting-package.R
\docType{package}
\name{sufficientForecasting-package}
\alias{sufficientForecasting}
\alias{sufficientForecasting-package}
\title{sufficientForecasting: Sufficient Forecasting using Factor Models}
\description{
This package implements the sufficient forecasting (SF) method for a single time series forecasting using many predictors and a possibly nonlinear forecasting function. Assuming that the predictors are driven by some latent factors, the SF first conducts factor analysis and then performs sufficient dimension reduction on the estimated factors to derive predictive indices for forecasting. The package implements several dimension reduction approaches, including principal components (PC), sliced inverse regression (SIR), and directional regression (DR).
}
\author{
\strong{Maintainer}: Jing Fu \email{jingfu991224@outlook.com}

Authors:
\itemize{
  \item Jianqing Fan
  \item Wei Luo
  \item Lingzhou Xue
  \item Jiawei Yao
  \item Xiufan Yu
}

}
\keyword{internal}
