% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sufficientForecasting-package.R
\docType{package}
\name{sufficientForecasting-package}
\alias{sufficientForecasting}
\alias{sufficientForecasting-package}
\title{sufficientForecasting: Sufficient Forecasting using Factor Models}
\description{
The sufficient forecasting (SF) method is implemented by this package for a single time series forecasting using many predictors and a possibly nonlinear forecasting function. Assuming that the predictors are driven by some latent factors, the SF first conducts factor analysis and then performs sufficient dimension reduction on the estimated factors to derive predictive indices for forecasting. The package implements several dimension reduction approaches, including principal components (PC), sliced inverse regression (SIR), and directional regression (DR).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JingFu1224/sufficientForecasting}
  \item Report bugs at \url{https://github.com/JingFu1224/sufficientForecasting/issues}
}

}
\author{
\strong{Maintainer}: Jing Fu \email{jingfu991224@outlook.com}

Authors:
\itemize{
  \item Jianqing Fan
  \item Wei Luo
  \item Lingzhou Xue
  \item Jiawei Yao
  \item Xiufan Yu
}

}
\keyword{internal}
